remove home's section login if a user is connected
